IPSec EMS/common/requirements_ipsecems.txt
Django==1.9.2
Jinja2==2.8
MarkupSafe==0.23
Yapsy==1.11.223
argparse==1.2.1
configparser==3.3.0r2
djangorestframework==3.3.2
ipaddress==1.0.16
netaddr==0.7.18
netifaces==0.10.4
prettytable==0.7.2
python-consul==0.4.7
requests==2.9.1
shell-command==0.1
six==1.10.0
wsgiref==0.1.2
